Consolidate Earn components with shared hook
